// lib/api/generated/models/workspaceMemberListResponseDataMembersItem.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */
import type { WorkspaceMemberListResponseDataMembersItemRole } from "./workspaceMemberListResponseDataMembersItemRole";

export type WorkspaceMemberListResponseDataMembersItem = {
  /**
   * 멤버 프로필 이미지 URL. 등록되지 않은 경우 null
   * @nullable
   */
  image: string | null;
  /** 워크스페이스 역할 */
  role: WorkspaceMemberListResponseDataMembersItemRole;
  /** 워크스페이스 가입 시각 */
  joinedAt: string;
  /** 멤버 이름 */
  name: string;
  /**
   * 멤버 사용자 식별자
   * @minLength 13
   * @maxLength 13
   * @pattern ^[0-9A-HJKMNP-TV-Z]{13}$
   */
  userId: string;
  /** 멤버 이메일 */
  email: string;
};